Added another special type to the configuration, refactored Dashboard component to be generic
