+++ src/app/components/IFrame/IFrame.jsx
@@ -1,9 +1,8 @@
 import React from 'react'
 import styled from 'styled-components'
-import { Column } from '@datapunt/asc-ui'
 
-import setIframeSize from '../../../../shared/services/set-iframe-size/setIframeSize'
-import LoadingIndicator from '../../../../shared/components/loading-indicator/LoadingIndicator'
+import setIframeSize from '../../../shared/services/set-iframe-size/setIframeSize'
+import LoadingIndicator from '../../../shared/components/loading-indicator/LoadingIndicator'
 
 const IFrameContainer = styled.div`
   position: relative;
@@ -15,7 +14,7 @@ const IFrameContainer = styled.div`
   }
 `
 
-const Dashboard = ({ contentLink, title }) => {
+const IFrame = ({ contentLink, title }) => {
   const [iframeLoading, setIframeLoading] = React.useState(true)
   const [iframeHeight, setIframeHeight] = React.useState(0)
   const iframeRef = React.useRef(null)
@@ -44,23 +43,21 @@ const Dashboard = ({ contentLink, title }) => {
   }, [iframeHeight])
 
   return (
-    <Column wrap span={{ small: 12, medium: 12, big: 12, large: 12, xLarge: 12 }}>
-      <IFrameContainer>
-        {iframeLoading && <LoadingIndicator />}
-        {contentLink && contentLink.uri && (
-          <iframe
-            src={contentLink.uri}
-            title={title}
-            ref={iframeRef}
-            onLoad={iframeLoaded}
-            width="100%"
-            height={iframeHeight}
-            frameBorder="0"
-          />
-        )}
-      </IFrameContainer>
-    </Column>
+    <IFrameContainer>
+      {iframeLoading && <LoadingIndicator />}
+      {contentLink && contentLink.uri && (
+        <iframe
+          src={contentLink.uri}
+          title={title}
+          ref={iframeRef}
+          onLoad={iframeLoaded}
+          width="100%"
+          height={iframeHeight}
+          frameBorder="0"
+        />
+      )}
+    </IFrameContainer>
   )
 }
 
-export default Dashboard
+export default IFrame
